add data gen to UiTestWindow
Specifically, in the WhatdidControlHooks panel. This moves the data gen
from ScreenshotGenerator to a new SampleData class in main whatdid (but
guarded by `#if UI_TEST`).
